client/krypton-python/pyproject.toml
[tool.poetry]
name = "krypton-python"
version = "0.1.0"
description = ""
authors = ["Maxime Mouchet <max@maxmouchet.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.1"
requests = "^2.23.0"
pyjwt = {extras = ["crypto"], version = "^1.7.1"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pylint = "^2.4.4"
pytest = "^5.4.1"
pre-commit = "^2.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
